+++ src/rain/entity/projectile/Projectile.java
@@ -6,7 +6,7 @@ import rain.graphics.Sprite;
 import java.util.Random;
 
 public abstract class Projectile extends Entity {
-    protected final int xOrigin, yOrigin;
+    protected final double xOrigin, yOrigin;
     protected double angle;
 
     protected double distance;
@@ -16,7 +16,7 @@ public abstract class Projectile extends Entity {
 
     protected final Random random = new Random();
 
-    public Projectile(int x, int y, double dir){
+    public Projectile(double x, double y, double dir){
         super(x,y,null);
         this.xOrigin = x;
         this.yOrigin = y;
